rename main_dev to main
online analysis features and bug fixes confirmed. make _dev version main
version.
(bug fixes: lagginess solved by changing save format, also fixed first
response plotting issue)
